requirements.txt
# Vast.ai SDK for PyWorker
# CRITICAL: Must be >= 0.3.1 for Worker, WorkerConfig, etc.
vastai-sdk>=0.3.1
